Extend the FlowApi from EngineApp
